seperate general preprocess functions and NSI specific functions to different files
