+++ mbirjax/__init__.py
@@ -8,3 +8,4 @@ from .vcd_utils import *
 from .memory_stats import *
 from .plot_utils import *
 from .preprocess import *
+from .preprocess_NSI import *
